# user-service/swagger.yaml
openapi: 3.0.3
info:
  title: User Service API
  description: |
    User Service for Smart Queue Management System
    
    This microservice manages user data including:
    - User registration and profile management
    - User authentication and authorization
    - User data retrieval and updates
    
    **Security Features:**
    - Environment-based MongoDB connection (no hardcoded credentials)
    - Input validation and sanitization
    - RESTful API design with proper HTTP status codes
    - Comprehensive error handling
  version: 1.0.0
  contact:
    name: Smart Queue Management Team
    email: team@smartqueue.local
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: https://user-service.prod.com
    description: Production server
paths:
...
  /api/users:
    post:
      summary: Create a new user
      description: Register a new user in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - phone
              properties:
                name:
                  type: string
                  description: User's full name
                  minLength: 2
                  maxLength: 100
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  description: User's email address (must be unique)
                  example: "john.doe@example.com"
                phone:
                  type: string
                  description: User's phone number in +94 format
                  pattern: '^\+94\d{9}$'
                  example: "+94771234567"
                nic:
                  type: string
                  description: User's National Identity Card number
                  example: "123456789V"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "507f1f77bcf86cd799439011"
                  name:
                    type: string
                    example: "John Doe"
                  email:
                    type: string
                    example: "john.doe@example.com"
                  phone:
                    type: string
                    example: "+94771234567"
                  created_at:
                    type: string
                    format: date-time
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User with this email already exists"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

    get:
      summary: Retrieve all users
      description: Get a list of all users with pagination support
      parameters:
        - name: page
          in: query
          description: Page number (starting from 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of users per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
                        phone:
                          type: string
                        nic:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []
...